requirements.txt
Django>=4.2
django-unfold
python-dotenv
django-import-export
pandas
openpyxl
